Revise command consequences API
