+++ client/src/main/kotlin/io/spine/chords/client/layout/CommandWizard.kt
@@ -30,7 +30,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
 import com.google.protobuf.Message
 import io.spine.base.CommandMessage
-import io.spine.chords.client.CommandConsequences
+import io.spine.chords.client.CommandConsequencesScope
 import io.spine.chords.client.form.CommandMessageForm
 import io.spine.chords.core.layout.AbstractWizardPage
 import io.spine.chords.core.layout.Wizard
@@ -73,7 +73,7 @@ public abstract class CommandWizard<C : CommandMessage, B : ValidatingBuilder<ou
             onBeforeBuild = { beforeBuild(it) }
         ) {
             validationDisplayMode = MANUAL
-            commandConsequences = ::commandConsequences
+            commandConsequences = { commandConsequences() }
             enabled = submitting
         }
 
@@ -98,20 +98,17 @@ public abstract class CommandWizard<C : CommandMessage, B : ValidatingBuilder<ou
     protected abstract fun createCommandBuilder(): B
 
     /**
-     * A function, which, given a command message that is about to be posted,
-     * should provide the [CommandConsequences] object that defines how the
-     * command's consequences should be handled.
+     * A function, which, should register handlers for consequences of
+     * command [C] posted by the wizard.
      *
-     * Note that the provided [CommandConsequences] instance has to be
-     * configured to perform any side effects that should follow the posting of
-     * the wizard's command. The typical minimum implementation would ensure
-     * that some event that is emitted after posting the command would lead to
-     * closing the wizard (by invoking the [close] method upon that event).
-     *
-     * @param command A command, which is going to be posted.
-     * @return A respectively configured [CommandConsequences] instance.
+     * The command, which is going to be posted and whose consequence handlers
+     * should be registered can be obtained from the
+     * [command][CommandConsequencesScope.command] property available in the
+     * function's scope, and handlers can be registered using the
+     * [`onXXX`][CommandConsequencesScope] functions available in the
+     * function's scope.
      */
-    protected abstract fun commandConsequences(command: C): CommandConsequences<C>
+    protected abstract fun CommandConsequencesScope<C>.commandConsequences()
 
     /**
      * Allows to programmatically amend the command message builder before
